Start testing
This is the first commit on our sample process. It makes sense to go
ahead and use a "real" process, we can make a separate branch, but
following a real process for testing/developing should be good!

This test obviously passes, but is also obviously not complete. There
are a number of other files that if we were being perfect we would add
in their own commit :P